+++ tests/features/test_smarterdb.py
@@ -14,12 +14,34 @@ from unittest.mock import patch
 from src.features.smarterdb import (
     VariantSheep, Location, SampleSheep,
     SmarterDBException, getSmarterId, Breed, get_or_create_breed, Dataset,
-    BreedAlias, get_or_create_sample, SEX, get_sample_type)
+    BreedAlias, get_or_create_sample, SEX, get_sample_type, Country)
 
 from ..common import MongoMockMixin, SmarterIDMixin
 
 # set data dir (like os.dirname(__file__)) + "fixtures"
 DATA_DIR = pathlib.Path(__file__).parents[1] / "fixtures"
+
+
+class CountryTestCase(MongoMockMixin, unittest.TestCase):
+    def setUp(self):
+        self.country = Country(name="Italy", species="Sheep")
+        self.country.save()
+
+    def tearDown(self):
+        Country.objects.delete()
+
+    def test_init(self):
+        self.assertEqual(self.country.alpha_2, "IT")
+        self.assertEqual(self.country.name, "Italy")
+        self.assertListEqual(self.country.species, ["Sheep"])
+
+    def test_str(self):
+        self.assertEqual(str(self.country), "Italy (IT)")
+
+    def test_no_official_name(self):
+        country = Country(name="Barbados")
+        self.assertEqual(str(country), "Barbados (BB)")
+        self.assertIsNone(country.official_name)
 
 
 class BreedTestCase(MongoMockMixin, unittest.TestCase):
@@ -510,6 +532,12 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
         self.alias = "TEST-ALIAS"
         self.type_ = "background"
 
+        # GPS coordinates
+        self.locations = {
+            'type': 'MultiPoint',
+            'coordinates': [[9.1859243, 45.4654219]]
+        }
+
     def tearDown(self):
         SampleSheep.objects().delete()
 
@@ -528,6 +556,9 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
         self.sample.country = self.country
         self.sample.breed = "Texel"
         self.sample.species = "Sheep"
+
+        # add locations
+        self.sample.locations = self.locations
 
         # save sample in db
         self.sample.save()
@@ -589,6 +620,47 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
 
         test = get_sample_type(self.dataset)
         self.assertEqual("background", test)
+
+    @unittest.skip(
+        "'$geoWithin' is a valid operation but it is "
+        "not supported by Mongomock yet")
+    def test_get_sample_geo_within(self):
+        self.create_sample()
+
+        qs = SampleSheep.objects.filter(
+            locations__geo_within={
+                "type": "Polygon",
+                "coordinates": [[
+                    [9, 45],
+                    [9, 46],
+                    [10, 46],
+                    [10, 45],
+                    [9, 45]
+                ]],
+            }
+        )
+
+        self.assertEqual(qs.count(), 1)
+        sample = qs.get()
+        self.assertEqual(sample.smarter_id, self.smarter_id)
+
+    @unittest.skip(
+        "'$geoWithin' is a valid operation but it is "
+        "not supported by Mongomock yet")
+    def test_get_sample_geo_within_sphere(self):
+        self.create_sample()
+
+        # center the sphere and define radius in radiant x / 6378.1 to convert
+        # kilometers to radias
+        qs = SampleSheep.objects.filter(
+            locations__geo_within_sphere=[
+                [9.18, 45.46],
+                10 / 6378.1]
+            )
+
+        self.assertEqual(qs.count(), 1)
+        sample = qs.get()
+        self.assertEqual(sample.smarter_id, self.smarter_id)
 
 
 class SEXTestCase(unittest.TestCase):