+++ HISTORY.rst
@@ -22,6 +22,17 @@ TODO
   both in *import_samples* and *import_metadata* scripts
 * define a collection for all available *purpose* phenotypes
 
+0.4.4.dev0
+----------
+
+* Model location with MultiPointField
+* Describe smarter metadata
+* Import sweden goat metadata
+* Import latest 290 samples greek dataset
+* Fix issue with greek samples name (``B273`` converted into ``B273A``)
+* add latest 19 sheep greek samples
+* add a country collection
+
 0.4.3 (2021-11-11)
 ------------------